refactor: 도메인 모델 패키지 구조 변경 및 author 하위 패키지 도입
- `Detail.kt` 및 `ListPage.kt`를 `model.author` 패키지로 이동
- `PlaylistDetail.kt` 및 `SearchedSong.kt`를 `model.author.playlist` 패키지로 이동
- 패키지 위치 변경에 따른 관련 클래스의 패키지 경로 및 임포트 업데이트
